ai-communication-expert/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# AI/LLM
anthropic==0.39.0
openai==1.54.3

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Document Generation
python-docx==1.1.2
reportlab==4.2.5
jinja2==3.1.4

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0

# Date handling
python-dateutil==2.9.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

